# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Travel Assistant API
  version: 1.0.0
  description: API per l'assistente virtuale di viaggi personalizzati.
servers:
  - url: http://localhost:4000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsBearer:
      type: http
      scheme: bearer
      bearerFormat: exporter-token
  schemas:
    SharedItinerary:
      type: object
      required: [id, details, totalCost, flightCost, hotelCost, activityCost, createdAt]
      properties:
        id: { type: string, format: uuid }
        details: { type: object, additionalProperties: true }
        totalCost: { type: number, format: float }
        flightCost: { type: number, format: float }
        hotelCost: { type: number, format: float }
        activityCost: { type: number, format: float }
        compromises:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
  /share/itineraries/{itineraryId}:
    get:
      summary: Recupera il link condivisibile di un itinerario dell'utente
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: itineraryId
          required: true
          description: Identificativo dell'itinerario
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Dati minimi dell'itinerario condivisibile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedItinerary'
        '403': { description: Itinerario non appartenente all'utente autenticato }
        '404': { description: Itinerario o link non trovato }
  /share/public/{token}:
    get:
      summary: Recupera un itinerario tramite token condiviso
      parameters:
        - in: path
          name: token
          required: true
          description: Token pubblico del link condiviso
          schema: { type: string }
      responses:
        '200':
          description: Dati minimi dell'itinerario condiviso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedItinerary'
        '403': { description: Link revocato o accesso non consentito }
        '404': { description: Token non trovato }
...
  /itineraries/{id}:
    get:
      summary: Recupera un itinerario
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Itinerario trovato }
        '404': { description: Itinerario non trovato }
  /itinerary-jobs:
    post:
      summary: Accoda una generazione itinerario idempotente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId, idempotencyKey]
              properties:
                conversationId: { type: string, format: uuid }
                idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '202': { description: Job creato e accodato }
        '200': { description: Job già esistente per la stessa chiave }
        '409': { description: Conversazione non pronta o chiave usata altrove }
  /itinerary-jobs/{id}:
    get:
      summary: Consulta stato, avanzamento, risultato o errore del job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Stato corrente del job }
        '404': { description: Job non trovato o appartenente a un altro utente }
  /bookings:
    post:
      summary: Conferma la prenotazione di un itinerario (idempotente)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itineraryId, idempotencyKey, confirm]
              properties:
                itineraryId: { type: string }
                idempotencyKey: { type: string, description: 'UUID generato lato client per prevenire duplicati' }
                confirm: { type: boolean, description: 'Deve essere true: conferma esplicita richiesta' }
      responses:
        '201': { description: Prenotazione creata }
        '200': { description: Prenotazione già processata (idempotenza) }
        '409': { description: Prenotazione fallita per indisponibilità sopravvenuta }
    get:
      summary: Lista prenotazioni dell'utente autenticato
      responses:
        '200': { description: Elenco prenotazioni }
  /bookings/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    delete:
      summary: Cancella una prenotazione confermata e ripristina le disponibilità
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: Prenotazione cancellata }
        '404': { description: Prenotazione non trovata o non appartenente all'utente }
        '409': { description: Stato non cancellabile o conflitto transazionale }
    patch:
      summary: Sostituisce il booking con un nuovo itinerario draft dell'utente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newItineraryId]
              properties:
                newItineraryId: { type: string }
      responses:
        '200': { description: Prenotazione modificata }
        '404': { description: Prenotazione o nuovo itinerario non trovato }
        '409': { description: Regola di modifica non rispettata o disponibilità insufficiente }